chore(helper):add handlePswrdComparison helper function to loginUser controller
